refactor(mapper): move the geolocation logic to separate component
